backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
flask-cors==4.0.0
alembic==1.13.0
